requirements.txt
Flask==1.1.1
Keras==2.4.3
Keras-Preprocessing==1.1.0
nltk==3.4.5
notebook==6.0.3
numpy==1.18.1
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
torch==1.7.0
torchvision==0.8.1
tqdm==4.42.1
Pillow==7.0.0
